POR 2740 - consolidate components (#276)
* reorganized file structure

* removed duplicate code for education, languages, and technologies components

* removed duplicate code in past experience components

* removed duplicate code in certifications components

* removed duplicate code in awards components

* moved EmployeeForm to form folder
