+++ src/components/employee-beta/cards/HireInfoCard.vue
