+++ src/components/employee-beta/cards/personal/OtherInfoCard.vue
@@ -1,10 +1,7 @@
 <template>
   <base-card title="Other Information">
-    <template #title>
-      <h3 class="text-white px-2">
-        <v-icon size="small" style="margin-right: 10px" id="personal" color="white"> mdi-information </v-icon>Other
-        Information
-      </h3>
+    <template #prependIcon>
+      <v-icon size="small" style="margin-right: 10px" id="personal" color="white"> mdi-information </v-icon>
     </template>
     <v-card-text class="px-7 pt-5 pb-1 text-black">
       <p v-if="!isEmpty(getAIN())"><b>AIN:</b> {{ getAIN() }}</p>
