+++ src/components/employee-beta/cards/personal/PersonalInfoCard.vue
@@ -1,10 +1,7 @@
 <template>
   <base-card title="Personal Information">
-    <template #title>
-      <h3 class="text-white px-2">
-        <v-icon size="small" style="margin-right: 10px" id="personal" color="white"> mdi-account </v-icon> Personal
-        Information
-      </h3>
+    <template #prependIcon>
+      <v-icon size="small" style="margin-right: 10px" id="personal" color="white"> mdi-account </v-icon>
     </template>
     <v-card-text class="px-7 pt-7 pb-1 text-black">
       <p v-if="!isEmpty(getPhoneNumber())">
