+++ src/components/employee-beta/modals/AwardsModal.vue
@@ -4,20 +4,7 @@
       <!-- Employee has Awards -->
       <div v-if="!isEmpty(filteredList)">
         <!-- Loop Awards -->
-        <v-list>
-          <v-list-item v-for="(awards, index) in filteredList" :key="awards.name + index">
-            <v-list-item-title class="d-flex align-center pb-4">
-              <v-icon class="mx-3">mdi-trophy-award</v-icon>
-              <p class="ma-3">
-                <b>{{ awards.name }}</b>
-              </p>
-            </v-list-item-title>
-            <v-row class="pl-10">
-              <p><b>Date Received: </b>{{ monthYearFormat(awards.dateReceived) }}</p>
-              <v-divider v-if="index < filteredList.length - 1" class="mb-3"></v-divider>
-            </v-row>
-          </v-list-item>
-        </v-list>
+        <awards-list :list="filteredList"></awards-list>
         <!-- End Loop Awards -->
       </div>
       <!-- Employee does not have Awards -->
@@ -32,8 +19,9 @@
 
 <script setup>
 import { ref, computed } from 'vue';
-import { isEmpty, monthYearFormat } from '@/utils/utils';
+import { isEmpty } from '@/utils/utils';
 import BaseInfoModal from '@/components/employee-beta/modals/BaseInfoModal.vue';
+import AwardsList from '../lists/AwardsList.vue';
 
 // |--------------------------------------------------|
 // |                                                  |
