+++ src/components/employee-beta/modals/CertificationsModal.vue
@@ -4,25 +4,7 @@
       <!-- Employee has Certifications -->
       <div v-if="!isEmpty(filteredList)">
         <!-- Loop Certifications -->
-        <v-list>
-          <v-list-item v-for="(certification, index) in filteredList" :key="certification.name + index">
-            <v-list-item-title class="d-flex align-center pb-4">
-              <v-icon class="mx-3">mdi-star-check</v-icon>
-              <p class="ma-3">
-                <b>{{ certification.name }}</b>
-              </p>
-            </v-list-item-title>
-            <v-row class="pl-10">
-              <p><b>Date Received: </b>{{ monthDayYearFormat(certification.dateReceived) }}</p>
-            </v-row>
-            <v-row class="pl-10">
-              <p v-if="certification.expirationDate">
-                <b>Expiration Date: </b>{{ monthDayYearFormat(certification.expirationDate) }}
-              </p>
-              <v-divider v-if="index < filteredList.length - 1" class="mb-3"></v-divider>
-            </v-row>
-          </v-list-item>
-        </v-list>
+        <certifications-list :list="filteredList"></certifications-list>
         <!-- End Loop Certifications -->
       </div>
       <!-- Employee does not have Certifications -->
@@ -40,8 +22,9 @@
 
 <script setup>
 import { ref, computed } from 'vue';
-import { isEmpty, monthDayYearFormat } from '@/utils/utils';
-import BaseInfoModal from '@/components/employee-beta/modals/BaseInfoModal.vue';
+import { isEmpty } from '@/utils/utils';
+import BaseInfoModal from '../modals/BaseInfoModal.vue';
+import CertificationsList from '../lists/CertificationsList.vue';
 
 // |--------------------------------------------------|
 // |                                                  |
